refactor(core): extract LinkPopover from Link mark into separate extension
